[render] rename qt/gui/editors path to qt/gui/toolbar_frames
