+++ vclib/render/include/vclib/qt/gui/toolbar_frames/axis_editor_frame.h
@@ -20,10 +20,54 @@
  * (https://www.mozilla.org/en-US/MPL/2.0/) for more details.                *
  ****************************************************************************/
 
-#ifndef VCL_QT_GUI_EDITORS_H
-#define VCL_QT_GUI_EDITORS_H
+#ifndef VCL_QT_GUI_TOOLBAR_FRAMES_AXIS_EDITOR_FRAME_H
+#define VCL_QT_GUI_TOOLBAR_FRAMES_AXIS_EDITOR_FRAME_H
 
-#include "editors/axis_editor_frame.h"
-#include "editors/bounding_box_editor_frame.h"
+#include "generic_editor_frame.h"
 
-#endif // VCL_QT_GUI_EDITORS_H
+#include <vclib/render/editors/axis_editor.h>
+
+namespace vcl::qt {
+
+template<typename ViewerType>
+class AxisEditorFrame : public GenericEditorFrame
+{
+    using Base = GenericEditorFrame;
+
+    std::shared_ptr<vcl::AxisEditor<ViewerType>> mAxisEditor;
+
+public:
+    explicit AxisEditorFrame(
+        std::shared_ptr<vcl::AxisEditor<ViewerType>> ptr,
+        QWidget*                                            parent = nullptr) :
+            GenericEditorFrame(parent)
+    {
+        mAxisEditor = ptr;
+
+        QIcon ic(":/icons/show_axis.png");
+
+        QPushButton* editorButton = Base::addButton(ic);
+
+        editorButton->setToolTip("Show Axis");
+
+        Base::hideSettingsButton();
+
+        mAxisEditor->setShortcutCallback([editorButton]() {
+            editorButton->click();
+        });
+
+        connect(
+            editorButton,
+            &QPushButton::clicked,
+            this,
+            [this]() {
+                if (mAxisEditor) {
+                    mAxisEditor->toggleVisibility();
+                }
+            });
+    }
+};
+
+} // namespace vcl::qt
+
+#endif // VCL_QT_GUI_TOOLBAR_FRAMES_AXIS_EDITOR_FRAME_H
