+++ vclib/render/src/vclib/qt/gui/toolbar_frames/generic_editor_frame.ui
