+++ vclib/render/src/vclib/qt/gui/toolbar_frames/settings/edit_mode_settings_frame.ui
